Fix: "수정, History entity", "수정, 대결 예정 목록에 대결 종료된 대결이 나오는 것을 수정", "수정, 히스토리 조회 시 최신순으로 조회되게 수정"

+++ src/main/java/me/coldrain/ninetyminute/service/TeamService.java
@@ -443,6 +443,7 @@ public class TeamService {
     public void disbandTeam(Long teamId, Member member) {
         if(teamRepository.findByIdAndDeletedTrue(teamId).orElse(null) != null)
             throw new IllegalArgumentException("팀을 찾을 수 없습니다.");
+        if (member.getOpenTeam() == null) throw new IllegalArgumentException("개설한 팀이 존재하지 않습니다.");
         if (member.getOpenTeam().getId().equals(teamId)) {
             teamRepository.deleteById(teamId);
             Member my = memberRepository.findById(member.getId()).orElseThrow(
